Rename session.rs to scripting_session.rs
