+++ crates/scripting_tool/src/scripting_session.rs
